pyproject.toml
[tool.poetry]
name = "infolibras"
version = "0.1.0"
description = ""
authors = ["Vitor Daniel <vitor036daniel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
Scrapy = "^2.11.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
